requirements.txt
flask==2.0.1
openai==1.3.0
pandas==2.0.0
openpyxl==3.1.2
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==20.1.0
